+++ crates/mimic/src/design/base/validator/text/iso.rs
@@ -1,4 +1,4 @@
-use crate::{core::traits::ValidatorString, design::prelude::*};
+use crate::{core::traits::Validator, design::prelude::*};
 
 ///
 /// Iso6391
@@ -10,10 +10,8 @@ use crate::{core::traits::ValidatorString, design::prelude::*};
 #[validator]
 pub struct Iso6391 {}
 
-impl ValidatorString for Iso6391 {
-    fn validate<S: AsRef<str>>(&self, s: S) -> Result<(), String> {
-        let s = s.as_ref();
-
+impl Validator<str> for Iso6391 {
+    fn validate(&self, s: &str) -> Result<(), String> {
         if s.len() != 2 || !s.chars().all(|c| c.is_ascii_lowercase()) {
             Err(format!("invalid ISO 3166-1 alpha-2 country code {s}"))
         } else {