+++ crates/mimic/src/design/mod.rs
@@ -14,8 +14,8 @@ pub mod prelude {
             Key, Value,
             traits::{
                 EntityFixture, EntityIdKind as _, EntityKind as _, NumCast, Path as _,
-                Serialize as _, TypeView as _, Validate as _, ValidateCustom, ValidatorBytes as _,
-                ValidatorDecimal as _, ValidatorNumber as _, ValidatorString as _, Visitable as _,
+                Serialize as _, TypeView as _, Validate as _, ValidateCustom, Validator as _,
+                Visitable as _,
             },
             types::*,
         },