+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - db::Db is now passed to fixtures and not the SaveExecutor
 - QueryShape::ALl removed because it's dangerous and doesn't add anything
 - Delete Range working
+- removed the Inner trait as to_view makes it obsolete
+- merged the ValidatorBytes, ValidatorString into just Validator
 
 ## [0.8.1] - Internal Changes
 - added CanisterKind, StoreKind, IndexKind so that we can declare the static types